ggExtra {ggExtra}	R Documentation

ggExtra

Description

Collection of functions and layers to enhance ggplot2. The main function is ggMarginal, which can be used to add marginal histograms/boxplots/density plots to ggplot2 scatterplots

Details

View a demo Shiny app or see the full README on GitHub.

[Package ggExtra version 0.8 Index]