rotateTextX {ggExtra}	R Documentation

Rotate x axis labels

Description

Rotate the labels on the x axis to be rotated so that they are vertical, which is often useful when there are many overlapping labels along the x axis.

Usage

rotateTextX(angle = 90, hjust = 1, vjust = 0.5)

Arguments

angle
Angle (in [0, 360])

hjust
Horizontal justification (in [0, 1])

vjust
Vertical justification (in [0, 1])

Details

This function is quite simple, but it can be useful if you don't have the exact syntax to do this engraved in your head.

Value

A ggplot2 layer that can be added to an existing ggplot2 object.

Examples

df <- data.frame(x = paste("Letter", LETTERS, sep = "_"),
                 y = seq_along(LETTERS))
p <- ggplot2::ggplot(df, ggplot2::aes(x, y)) + ggplot2::geom_point()
p + rotateTextX()

[Package ggExtra version 0.8 Index]